+++ packages/helpers/src/index.ts
@@ -2,6 +2,7 @@ export { default as bound } from "./bound";
 export { default as createMap } from "./createMap";
 export * from "./disposable";
 export * from "./functionHelpers";
+export { default as groupBy } from "./groupBy";
 export { default as ManualPromise } from "./manualPromise";
 export * from "./nameof";
 export * from "./observableHelpers";